@@ -8,16 +8,20 @@
     "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
     "antd": "^4.24.16",
-    "apexcharts": "^3.27.1",
+    "apexcharts": "^4.5.0",
     "axios": "^1.7.7",
     "cross-env": "^7.0.3",
+    "date-fns": "^3.0.0",
     "install": "^0.13.0",
     "postcss": "^8.4.47",
-    "react": "^17.0.2",
-    "react-apexcharts": "^1.3.9",
-    "react-dom": "^17.0.2",
+    "react": "^19.1.0",
+    "react-apexcharts": "^1.7.0",
+    "react-date-range": "^2.0.1",
+    "react-datepicker": "^8.3.0",
+    "react-dom": "^19.1.0",
     "react-router-dom": "^5.2.0",
     "react-scripts": "^5.0.1",
+    "recharts": "^2.15.3",
     "resolve-url-loader": "^5.0.0",
     "styled-components": "^5.3.0",
     "web-vitals": "^1.1.2",
