@@ -1,13 +1,12 @@
-
 import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";  // Updated import
 import { BrowserRouter } from "react-router-dom";
 import App from "./App";
 
-
-ReactDOM.render(
+// Create a root and render
+const root = createRoot(document.getElementById("root"));
+root.render(
   <BrowserRouter>
     <App />
-  </BrowserRouter>,
-  document.getElementById("root"),
+  </BrowserRouter>
 );
