@@ -34,13 +34,13 @@ function Home() {
             Authorization: `Bearer ${token}`,
           },
         });
-
-        // Check if the response status is 401 (Unauthorized)
-        // if (response.status === 401) {
-        //   message.error("Session Expired.");
-        //   localStorage.removeItem("token"); // Remove the token from local storage
-        //   window.location.href = "/sign-in"; // Redirect to the login page
-        // }
+        console.log("cek error",response.status );
+       // Check if the response status is 401 (Unauthorized)
+        if (response.status === 401) {
+          message.error("Session Expired.");
+          localStorage.removeItem("token"); // Remove the token from local storage
+          window.location.href = "/sign-in"; // Redirect to the login page
+        }
 
         const result = await response.json();
         var total = "0";
@@ -73,6 +73,7 @@ function Home() {
           done: done,
         });
       } catch (error) {
+       
         console.error("Error fetching counts:", error);
       }
     };
